Add Magic Link SDK
