+++ Enum/AuthMethod.swift
@@ -1,0 +1,13 @@
+//
+//  AuthMethod.swift
+//  Cotter
+//
+//  Created by Albert Purnama on 6/27/20.
+//
+
+import UIKit
+
+public enum AuthMethod: String {
+    case magicLink = "MAGIC_LINK"
+    case otp = "OTP"
+}
