+++ Example/CotterIOS/AppDelegate.swift
@@ -41,10 +41,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         Cotter.configureWithLaunchOptions(launchOptions: launchOptions,apiSecretKey: apiSecretKey, apiKeyID: apiKeyID)
 
         // if you want to start PINViewControler on startup, use getPINViewController
-        let vc = CotterWrapper.cotter!.getPINViewController(hideClose: true, cb: Callback.shared.authCb)
-        let nav = UINavigationController(rootViewController: vc)
-        self.window?.rootViewController = nav
-        self.window?.makeKeyAndVisible()
+        // let vc = CotterWrapper.cotter!.getPINViewController(hideClose: true, cb: Callback.shared.authCb)
+        // let nav = UINavigationController(rootViewController: vc)
+        // self.window?.rootViewController = nav
+        // self.window?.makeKeyAndVisible()
         
         UNUserNotificationCenter.current() // 1
         .requestAuthorization(options: [.alert, .sound, .badge]) { // 2
